Working on Testing
Tried to debug full stack and wasn't maing progress. New approach is to
test the Service layer and consequently the Dao layer by creating
Service Layer tests. Much easier to debug and can gain confidence that
the back end (minus the controllers) is functioning correctly. With that
confidence, it will be easier to ensure the controllers and lastly UI is
working correctly.